Fix compilation ambiguity for system jsoncpp
In upstream JsonCpp there are two [] operators for Json::Value - one taking unsigned int, second taking int - passing long int leads to ambiguity.

+++ src/KeyFrame.cpp
@@ -367,7 +367,7 @@ void Keyframe::SetJsonValue(Json::Value root) {
 
 	if (!root["Points"].isNull())
 		// loop through points
-		for (long int x = 0; x < root["Points"].size(); x++) {
+		for (unsigned int x = 0; x < root["Points"].size(); x++) {
 			// Get each point
 			Json::Value existing_point = root["Points"][x];
